Removed the older query_builder and added new code base
