+++ src/databricks/labs/remorph/config.py
@@ -3,9 +3,7 @@ from dataclasses import dataclass
 
 from sqlglot.dialects.dialect import Dialect, Dialects
 
-from databricks.labs.remorph.reconcile.query_builder_refactored.recon_config import (
-    Table,
-)
+from databricks.labs.remorph.reconcile.recon_config import Table
 from databricks.labs.remorph.snow import databricks, experimental, snowflake
 
 logger = logging.getLogger(__name__)
