+++ tests/unit/reconcile/query_builder/test_sampling_query.py
@@ -1,13 +1,13 @@
 from pyspark.sql.types import IntegerType, StringType, StructField, StructType
 
-from databricks.labs.remorph.reconcile.query_builder_refactored.recon_config import (
+from databricks.labs.remorph.reconcile.query_builder.sampling_query import (
+    SamplingQueryBuilder,
+)
+from databricks.labs.remorph.reconcile.recon_config import (
     ColumnMapping,
     Filters,
     Schema,
     Transformation,
-)
-from databricks.labs.remorph.reconcile.query_builder_refactored.sampling_query import (
-    SamplingQueryBuilder,
 )
